총 수익금을 계산한다.

+++ src/main/kotlin/lotto/service/LottoCalculator.kt
@@ -10,7 +10,7 @@ class LottoCalculator {
                 .size
         }
 
-        fun calculatePrizeMoney(matchCount: Int): Int {
+        fun calculatePrizeMoney(matchCount: Int): Long {
             return when (matchCount) {
                 3 -> 5_000
                 4 -> 50_000
